+++ build.gradle
@@ -49,7 +49,8 @@ dependencies {
     compile group: 'org.apache.logging.log4j', name: 'log4j-core', version: '2.1'
     compile group: 'org.apache.commons', name: 'commons-lang3', version: '3.3.2'
     compile group: 'org.jsoup', name: 'jsoup', version: '1.9.2'
-    compile group: 'com.ibm.watson.developer_cloud', name: 'java-sdk', version:'3.5.3'
+    compile group: 'com.ibm.watson.developer_cloud', name: 'conversation', version:'3.5.3'
+    compile group: 'com.ibm.watson.developer_cloud', name: 'discovery', version:'3.5.3'
     testCompile 'com.sun.jersey:jersey-client:1.19.1'
     testCompile 'org.mockito:mockito-core:1.+'
     testCompile 'org.powermock:powermock-module-junit4:1.6.5'
@@ -75,7 +76,7 @@ artifacts {
 }
 
 war {
-    archiveName = 'conversation-enhanced-0.1-SNAPSHOT.war'
+    archiveName = 'conversation-with-discovery-0.1-SNAPSHOT.war'
 }
 
 gulp_build.dependsOn 'installGulp'
@@ -107,6 +108,13 @@ gulp {
 jacocoTestReport {
     reports {
         xml.enabled true
-        html.enabled false
+        html.enabled true
     }
+}
+
+tasks.withType(FindBugs) {
+  reports {
+      xml.enabled = false
+      html.enabled = true
+  }
 }
