+++ gradle.properties
@@ -1,3 +1,3 @@
 group   = com.ibm.watson.wea.app
-version = 0.1-SNAPSHOT
-appname = conversation-enhanced
+version = 0.5.0-SNAPSHOT
+appname = conversation-with-discovery
