+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'conversation-enhanced'
+rootProject.name = 'conversation-with-discovery'
