+++ src/main/java/com/ibm/watson/apis/conversation_enhanced/listener/SetupThread.java
@@ -1,15 +1,14 @@
-/**
- * (C) Copyright IBM Corp. 2016. All Rights Reserved.
- * 
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License. You may obtain a copy of the License at
- * 
+/*
+ * Copyright 2015 IBM Corp. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
  * http://www.apache.org/licenses/LICENSE-2.0
- * 
- * Unless required by applicable law or agreed to in writing, software distributed under the License
- * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
- * or implied. See the License for the specific language governing permissions and limitations under
- * the License.
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
  */
 package com.ibm.watson.apis.conversation_enhanced.listener;
 
@@ -22,8 +21,8 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 import com.google.gson.JsonObject;
-import com.ibm.watson.apis.conversation_enhanced.utils.Constants;
-import com.ibm.watson.apis.conversation_enhanced.utils.Messages;
+import com.ibm.watson.apis.conversation_with_discovery.utils.Constants;
+import com.ibm.watson.apis.conversation_with_discovery.utils.Messages;
 import com.ibm.watson.developer_cloud.conversation.v1.ConversationService;
 import com.ibm.watson.developer_cloud.conversation.v1.model.MessageRequest;
 import com.ibm.watson.developer_cloud.discovery.v1.Discovery;
@@ -31,9 +30,8 @@ import com.ibm.watson.developer_cloud.discovery.v1.model.query.QueryRequest;
 import com.ibm.watson.developer_cloud.service.exception.UnauthorizedException;
 
 /**
- * This class is forked when the application is accessed for the first time. It
- * tests if the conversation and discovery service associated with the app
- * (either bound on bluemix or specified in server.env)
+ * This class is forked when the application is accessed for the first time. It tests if the conversation and discovery
+ * service associated with the app (either bound on bluemix or specified in server.env)
  */
 public class SetupThread extends Thread {
 
@@ -41,8 +39,55 @@ public class SetupThread extends Thread {
 
   private List<PropertyChangeListener> listener = new ArrayList<PropertyChangeListener>();
 
+  /** The config. */
   public JsonObject config = new JsonObject();
 
+  /**
+   * Method to update the listeners about any property changes. This is used by the UI to inform user what the status of
+   * the setup.
+   *
+   * @param object
+   * @param config
+   */
+  private void notifyListeners(Object object, JsonObject config) {
+    for (PropertyChangeListener configUpdate : listener) {
+      configUpdate.propertyChange(new PropertyChangeEvent(this, "configSetup", config, config)); //$NON-NLS-1$
+    }
+  }
+
+  /**
+   * This method is used to update the JSON Config Object that will be sent back to the UI.
+   *
+   * @param config
+   * @param setupStep The step at which the setup is
+   * @param setupState The state of the setup(accepts ready/not_ready)
+   * @param setupPhase The current phase of the setup(this can be that the services is being setup or if any other
+   *        phase of the application is being setup)
+   * @param setupMessage The message that you want to be shown in the UI.
+   */
+  private void updateConfigObject(String setupStep, String setupState, String setupPhase, String setupMessage) {
+    config.addProperty(Constants.SETUP_STEP, setupStep); // $NON-NLS-1$
+    config.addProperty(Constants.SETUP_STATE, setupState); // $NON-NLS-1$
+    config.addProperty(Constants.SETUP_PHASE, setupPhase); // $NON-NLS-1$
+    config.addProperty(Constants.SETUP_MESSAGE, setupMessage); // $NON-NLS-1$
+    notifyListeners(this, config);
+  }
+
+  /**
+   * Method to add a listener.
+   *
+   * @param newListener PropertyChangeListener
+   */
+  public void addChangeListener(PropertyChangeListener newListener) {
+    listener.add(newListener);
+  }
+
+  /*
+   * (non-Javadoc)
+   *
+   * @see java.lang.Thread#run()
+   */
+  @Override
   public void run() {
 
     String status = "";
@@ -59,11 +104,11 @@ public class SetupThread extends Thread {
       String collectionId = System.getenv("DISCOVERY_COLLECTION_ID");
       String environmentId = System.getenv("DISCOVERY_ENVIRONMENT_ID");
 
-      if (userName == null || password == null || collectionId == null || environmentId == null) {
+      if ((userName == null) || (password == null) || (collectionId == null) || (environmentId == null)) {
         throw new IllegalArgumentException(Messages.getString("SetupThread.DISC_INVALID_CREDS"));
       }
-      if (userName.length() == 0 || password.length() == 0 || collectionId.length() == 0
-          || environmentId.length() == 0) {
+      if ((userName.length() == 0) || (password.length() == 0) || (collectionId.length() == 0)
+          || (environmentId.length() == 0)) {
         throw new IllegalArgumentException(Messages.getString("SetupThread.DISC_INVALID_CREDS"));
       }
 
@@ -86,10 +131,10 @@ public class SetupThread extends Thread {
       password = System.getenv("CONVERSATION_PASSWORD");
       String workspaceId = System.getenv("WORKSPACE_ID");
 
-      if (userName == null || password == null || workspaceId == null) {
+      if ((userName == null) || (password == null) || (workspaceId == null)) {
         throw new IllegalArgumentException(Messages.getString("SetupThread.CONV_INVALID_CREDS"));
       }
-      if (userName.length() == 0 || password.length() == 0 || workspaceId.length() == 0) {
+      if ((userName.length() == 0) || (password.length() == 0) || (workspaceId.length() == 0)) {
         throw new IllegalArgumentException(Messages.getString("SetupThread.CONV_INVALID_CREDS"));
       }
 
@@ -113,52 +158,5 @@ public class SetupThread extends Thread {
             e.getMessage() + " " + Messages.getString("SetupThread.CHECK_LOGS"));
       }
     }
-  }
-
-  /**
-   * Method to update the listeners about any property changes. This is used by
-   * the UI to inform user what the status of the setup.
-   * 
-   * @param object
-   * @param config
-   */
-  private void notifyListeners(Object object, JsonObject config) {
-    for (PropertyChangeListener configUpdate : listener) {
-      configUpdate.propertyChange(new PropertyChangeEvent(this, "configSetup", config, config)); //$NON-NLS-1$
-    }
-  }
-
-  /**
-   * Method to add a listener
-   * 
-   * @param newListener
-   *          PropertyChangeListener
-   */
-  public void addChangeListener(PropertyChangeListener newListener) {
-    listener.add(newListener);
-  }
-
-  /**
-   * This method is used to update the JSON Config Object that will be sent back
-   * to the UI.
-   * 
-   * @param config
-   * @param setup_step
-   *          The step at which the setup is
-   * @param setup_state
-   *          The state of the setup(accepts ready/not_ready)
-   * @param setup_phase
-   *          The current phase of the setup(this can be that the services is
-   *          being setup or if any other phase of the application is being
-   *          setup)
-   * @param setup_message
-   *          The message that you want to be shown in the UI.
-   */
-  private void updateConfigObject(String setup_step, String setup_state, String setup_phase, String setup_message) {
-    config.addProperty(Constants.SETUP_STEP, setup_step); // $NON-NLS-1$
-    config.addProperty(Constants.SETUP_STATE, setup_state); // $NON-NLS-1$
-    config.addProperty(Constants.SETUP_PHASE, setup_phase); // $NON-NLS-1$
-    config.addProperty(Constants.SETUP_MESSAGE, setup_message); // $NON-NLS-1$
-    notifyListeners(this, config);
   }
 }
