--- src/main/java/com/ibm/watson/apis/conversation_enhanced/utils/Messages.java
@@ -1,42 +1,0 @@
-/**
- * (C) Copyright IBM Corp. 2016. All Rights Reserved.
- * 
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License. You may obtain a copy of the License at
- * 
- * http://www.apache.org/licenses/LICENSE-2.0
- * 
- * Unless required by applicable law or agreed to in writing, software distributed under the License
- * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
- * or implied. See the License for the specific language governing permissions and limitations under
- * the License.
- */
-package com.ibm.watson.apis.conversation_enhanced.utils;
-
-import java.util.MissingResourceException;
-import java.util.ResourceBundle;
-
-/**
- * 
- * 
- * @author dgterry
- * 
- * 
- *         Jun 21, 2016
- */
-public class Messages {
-  private static final String BUNDLE_NAME = "locale/messages"; //$NON-NLS-1$
-
-  private static final ResourceBundle RESOURCE_BUNDLE = ResourceBundle.getBundle(BUNDLE_NAME);
-
-  private Messages() {
-  }
-
-  public static String getString(String key) {
-    try {
-      return RESOURCE_BUNDLE.getString(key);
-    } catch (MissingResourceException e) {
-      return '!' + key + '!';
-    }
-  }
-}
