+++ src/main/java/com/ibm/watson/apis/conversation_with_discovery/rest/ProxyResource.java
@@ -1,18 +1,17 @@
-/**
- * Copyright IBM Corp. 2016
+/*
+ * Copyright 2015 IBM Corp. All Rights Reserved.
  *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License. You may obtain a copy of the License at
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
  *
  * http://www.apache.org/licenses/LICENSE-2.0
  *
- * Unless required by applicable law or agreed to in writing, software distributed under the License
- * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
- * or implied. See the License for the specific language governing permissions and limitations under
- * the License.
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
  */
 
-package com.ibm.watson.apis.conversation_enhanced.rest;
+package com.ibm.watson.apis.conversation_with_discovery.rest;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -35,41 +34,38 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 import com.google.gson.Gson;
-import com.ibm.watson.apis.conversation_enhanced.discovery.DiscoveryClient;
-import com.ibm.watson.apis.conversation_enhanced.payload.DocumentPayload;
-import com.ibm.watson.apis.conversation_enhanced.utils.Constants;
-import com.ibm.watson.apis.conversation_enhanced.utils.Messages;
+import com.ibm.watson.apis.conversation_with_discovery.discovery.DiscoveryClient;
+import com.ibm.watson.apis.conversation_with_discovery.payload.DocumentPayload;
+import com.ibm.watson.apis.conversation_with_discovery.utils.Constants;
+import com.ibm.watson.apis.conversation_with_discovery.utils.Messages;
 import com.ibm.watson.developer_cloud.conversation.v1.ConversationService;
 import com.ibm.watson.developer_cloud.conversation.v1.model.MessageRequest;
 import com.ibm.watson.developer_cloud.conversation.v1.model.MessageResponse;
 import com.ibm.watson.developer_cloud.service.exception.UnauthorizedException;
 import com.ibm.watson.developer_cloud.util.GsonSingleton;
 
+/**
+ * The Class ProxyResource.
+ */
 @Path("conversation/api/v1/workspaces")
 public class ProxyResource {
+  private static String API_VERSION;
+  private static final String ERROR = "error";
   private static final Logger logger = LogManager.getLogger(ProxyResource.class.getName());
 
-  private static String API_VERSION;
-
-  private static final String ERROR = "error";
-
-  private String password = System.getenv("CONVERSATION_PASSWORD");
-
-  private String url;
-
-  private String username = System.getenv("CONVERSATION_USERNAME");
-
-  private DiscoveryClient discoveryClient = new DiscoveryClient();
-
+  /**
+   * Sets the conversation API version.
+   *
+   * @param version the new conversation API version
+   */
   public static void setConversationAPIVersion(String version) {
     API_VERSION = version;
   }
 
-  public void setCredentials(String username, String password, String url) {
-    this.username = username;
-    this.password = password;
-    this.url = url;
-  }
+  private DiscoveryClient discoveryClient = new DiscoveryClient();
+  private String password = System.getenv("CONVERSATION_PASSWORD");
+  private String url;
+  private String username = System.getenv("CONVERSATION_USERNAME");
 
   private MessageRequest buildMessageFromPayload(InputStream body) {
     StringBuilder sbuilder = null;
@@ -90,7 +86,9 @@ public class ProxyResource {
       logger.error(Messages.getString("ProxyResource.JSON_READ"), e);
     } finally {
       try {
-        reader.close();
+        if (reader != null) {
+          reader.close();
+        }
       } catch (IOException e) {
         logger.error(Messages.getString("ProxyResource.STREAM_CLOSE"), e);
       }
@@ -99,19 +97,15 @@ public class ProxyResource {
   }
 
   /**
-   * This method is responsible for sending the query the user types into the UI
-   * to the Watson services. The code demonstrates how the conversation service
-   * is called, how the response is evaluated, and how the response is then sent
-   * to the discovery service if necessary.
-   * 
-   * @param request
-   *          The full query the user asked of Watson
-   * @param id
-   *          The ID of the conversational workspace
-   * @return The response from Watson. The response will always contain the
-   *         conversation service's response. If the intent confidence is high
-   *         or the intent is out_of_scope, the response will also contain
-   *         information from the discovery service
+   * This method is responsible for sending the query the user types into the UI to the Watson services. The code
+   * demonstrates how the conversation service is called, how the response is evaluated, and how the response is then
+   * sent to the discovery service if necessary.
+   *
+   * @param request The full query the user asked of Watson
+   * @param id The ID of the conversational workspace
+   * @return The response from Watson. The response will always contain the conversation service's response. If the
+   *         intent confidence is high or the intent is out_of_scope, the response will also contain information from
+   *         the discovery service
    */
   private MessageResponse getWatsonResponse(MessageRequest request, String id) throws Exception {
 
@@ -119,9 +113,9 @@ public class ProxyResource {
     // appropriate conversation
     // service. Specific information is obtained from the VCAP_SERVICES
     // environment variable
-    ConversationService service = new ConversationService(
-        API_VERSION != null ? API_VERSION : ConversationService.VERSION_DATE_2016_09_20);
-    if (username != null || password != null) {
+    ConversationService service =
+        new ConversationService(API_VERSION != null ? API_VERSION : ConversationService.VERSION_DATE_2016_09_20);
+    if ((username != null) || (password != null)) {
       service.setUsernameAndPassword(username, password);
     }
 
@@ -136,12 +130,12 @@ public class ProxyResource {
     // should call the discovery service to obtain better answers
 
     if (response.getOutput().containsKey("action")
-        && response.getOutput().get("action").toString().indexOf("call_discovery") != -1) {
+        && (response.getOutput().get("action").toString().indexOf("call_discovery") != -1)) {
       String query = response.getInputText();
 
       // Extract the user's original query from the conversational
       // response
-      if (query != null && !query.isEmpty()) {
+      if ((query != null) && !query.isEmpty()) {
 
         // For this app, both the original conversation response and the
         // discovery response
@@ -170,6 +164,13 @@ public class ProxyResource {
     return response;
   }
 
+  /**
+   * Post message.
+   *
+   * @param id the id
+   * @param body the body
+   * @return the response
+   */
   @POST
   @Path("{id}/message")
   @Consumes(MediaType.APPLICATION_JSON)
@@ -205,5 +206,18 @@ public class ProxyResource {
       return Response.ok(new Gson().toJson(errorsOutput, HashMap.class)).type(MediaType.APPLICATION_JSON).build();
     }
     return Response.ok(new Gson().toJson(response, MessageResponse.class)).type(MediaType.APPLICATION_JSON).build();
+  }
+
+  /**
+   * Sets the credentials.
+   *
+   * @param username the username
+   * @param password the password
+   * @param url the url
+   */
+  public void setCredentials(String username, String password, String url) {
+    this.username = username;
+    this.password = password;
+    this.url = url;
   }
 }
