+++ src/main/java/com/ibm/watson/apis/conversation_with_discovery/rest/SetupResource.java
@@ -1,17 +1,16 @@
-/**
- * (C) Copyright IBM Corp. 2016. All Rights Reserved.
- * 
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License. You may obtain a copy of the License at
- * 
+/*
+ * Copyright 2015 IBM Corp. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
  * http://www.apache.org/licenses/LICENSE-2.0
- * 
- * Unless required by applicable law or agreed to in writing, software distributed under the License
- * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
- * or implied. See the License for the specific language governing permissions and limitations under
- * the License.
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
  */
-package com.ibm.watson.apis.conversation_enhanced.rest;
+package com.ibm.watson.apis.conversation_with_discovery.rest;
 
 import java.text.MessageFormat;
 
@@ -26,15 +25,14 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 import com.google.gson.JsonObject;
-import com.ibm.watson.apis.conversation_enhanced.listener.ServletContextListener;
-import com.ibm.watson.apis.conversation_enhanced.utils.Constants;
-import com.ibm.watson.apis.conversation_enhanced.utils.Messages;
+import com.ibm.watson.apis.conversation_enhanced.listener.AppServletContextListener;
+import com.ibm.watson.apis.conversation_with_discovery.utils.Constants;
+import com.ibm.watson.apis.conversation_with_discovery.utils.Messages;
 
 /**
- * The SetupResource is used to fetch the configuration file with the
- * WorkspaceId from the env file and the setup configuration stage that the
- * application is currently in. The API endpoint points to '/rest/setup'
- * 
+ * The SetupResource is used to fetch the configuration file with the WorkspaceId from the env file and the setup
+ * configuration stage that the application is currently in. The API endpoint points to '/rest/setup'
+ *
  */
 @Path("setup")
 public class SetupResource {
@@ -42,14 +40,14 @@ public class SetupResource {
 
   /**
    * Method to fetch config JSON object and also the workspace_id.
-   * 
+   *
    * @return response
    */
   @GET
   @Produces(MediaType.APPLICATION_JSON)
   public Response getConfig() {
-    String workspace_id = System.getenv(Constants.WORKSPACE_ID);
-    logger.debug(MessageFormat.format(Messages.getString("SetupResource.WORKSPACE_ID_IS"), workspace_id));
+    String workspaceId = System.getenv(Constants.WORKSPACE_ID);
+    logger.debug(MessageFormat.format(Messages.getString("SetupResource.WORKSPACE_ID_IS"), workspaceId));
 
     JsonObject config = new JsonObject();
     config.addProperty(Constants.SETUP_STATUS_MESSAGE, Messages.getString("SetupResource.SETUP_STATUS_MSG"));
@@ -59,14 +57,14 @@ public class SetupResource {
     config.addProperty(Constants.SETUP_MESSAGE, Messages.getString("SetupResource.CHECK_LOGS"));
 
     // Fetch the updated config JSON object from the servlet listener
-    config = new ServletContextListener().getJsonConfig();
+    config = new AppServletContextListener().getJsonConfig();
     config.addProperty(Constants.SETUP_STATUS_MESSAGE, Messages.getString("SetupResource.SETUP_STATUS_MSG"));
     logger.debug(Messages.getString("SetupResource.CONFIG_STATUS") + config);
 
-    if (config.get(Constants.SETUP_STEP).getAsInt() == 3
+    if ((config.get(Constants.SETUP_STEP).getAsInt() == 3)
         && config.get(Constants.SETUP_STATE).getAsString().equalsIgnoreCase(Constants.READY)) {
-      if (StringUtils.isNotBlank(workspace_id)) {
-        config.addProperty(Constants.WORKSPACE_ID, workspace_id);
+      if (StringUtils.isNotBlank(workspaceId)) {
+        config.addProperty(Constants.WORKSPACE_ID, workspaceId);
       } else {
         config.addProperty(Constants.SETUP_STEP, "0");
         config.addProperty(Constants.SETUP_STATE, Constants.NOT_READY);
