+++ src/main/java/com/ibm/watson/apis/conversation_with_discovery/rest/WinkApplication.java
@@ -1,17 +1,16 @@
-/**
- * (C) Copyright IBM Corp. 2016. All Rights Reserved.
- * 
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License. You may obtain a copy of the License at
- * 
+/*
+ * Copyright 2015 IBM Corp. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
  * http://www.apache.org/licenses/LICENSE-2.0
- * 
- * Unless required by applicable law or agreed to in writing, software distributed under the License
- * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
- * or implied. See the License for the specific language governing permissions and limitations under
- * the License.
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
  */
-package com.ibm.watson.apis.conversation_enhanced.rest;
+package com.ibm.watson.apis.conversation_with_discovery.rest;
 
 import java.util.HashSet;
 import java.util.Set;
@@ -19,8 +18,9 @@ import java.util.Set;
 import javax.ws.rs.ApplicationPath;
 import javax.ws.rs.core.Application;
 
-import com.ibm.watson.apis.conversation_enhanced.rest.ProxyResource;
-
+/**
+ * The Class WinkApplication.
+ */
 @ApplicationPath("/rest")
 /**
  * The Wink application which is used to add endpoints to the server.
