+++ src/main/resources/workspace.json
@@ -1994,7 +1994,7 @@
         },
         "text": {
           "values": [
-            "That question is out of scope for this application, take a look at the Conversation Enhanced application to handle questions like these."
+            "That question is out of scope for this application, take a look at the Conversation with Discovery application to handle questions like these."
           ]
         }
       },
