+++ src/main/webapp/WEB-INF/web.xml
@@ -3,7 +3,7 @@
 	xmlns="http://java.sun.com/xml/ns/javaee"
 	xsi:schemaLocation="http://java.sun.com/xml/ns/javaee http://java.sun.com/xml/ns/javaee/web-app_3_0.xsd"
 	version="3.0">
-	<display-name>Conversation Enhanced Demo</display-name>
+	<display-name>Conversation with Discovery Demo</display-name>
 	<welcome-file-list>
 		<welcome-file>redirect.html</welcome-file>
 	</welcome-file-list>
