+++ src/main/webapp/index.html
@@ -1,6 +1,6 @@
 <html>
 <head>
-  <title>Conversation Enhanced Chat App</title>
+  <title>Conversation with Discovery Chat App</title>
   <meta name="viewport" content="width=device-width, initial-scale=1">
   <!-- Load libraries -->
   <!-- IE required polyfills, in this exact order -->
@@ -31,8 +31,8 @@
   <div class="active-circle"></div>
   <div class="non-active-circle"></div>
   </div>
-  <a href="https://github.com/watson-developer-cloud/conversation-enhanced#import-a-workspace" target="_blank" class="werror">Go to Github </a>
-  <a href="https://github.com/watson-developer-cloud/conversation-enhanced#troubleshooting-in-bluemix" target="_blank" class="gerror">Go to Github </a>
+  <a href="https://github.com/watson-developer-cloud/conversation-with-discovery#import-a-workspace" target="_blank" class="werror">Go to Github </a>
+  <a href="https://github.com/watson-developer-cloud/conversation-with-discovery#troubleshooting-in-bluemix" target="_blank" class="gerror">Go to Github </a>
 </div>
 
 <chat-app>Loading...</chat-app>
