+++ src/test/java/com/ibm/watson/apis/conversation_with_discovery/rest/ProxyResourceTest.java
@@ -1,4 +1,16 @@
-package com.ibm.watson.apis.conversation_enhanced.rest;
+/*
+ * Copyright 2015 IBM Corp. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+package com.ibm.watson.apis.conversation_with_discovery.rest;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
@@ -33,19 +45,28 @@ import okhttp3.mockwebserver.MockWebServer;
 import okhttp3.mockwebserver.RecordedRequest;
 
 /**
- * Unit tests for the {@link ConversationService}
+ * Unit tests for the {@link ConversationService}.
  */
 
 public class ProxyResourceTest {
   private static final String FIXTURE = "src/test/resources/conversation.json";
   private static final String WORKSPACE_ID = "123";
+
+  /** The server. */
   protected MockWebServer server;
+
+  /** The Constant CONTENT_TYPE. */
   protected static final String CONTENT_TYPE = "Content-Type";
 
   /*
    * (non-Javadoc)
-   * 
+   *
    * @see com.ibm.watson.developer_cloud.WatsonServiceTest#setUp()
+   */
+  /**
+   * Sets the up.
+   *
+   * @throws Exception the exception
    */
   // @Override
   @Before
@@ -54,6 +75,11 @@ public class ProxyResourceTest {
     server.start();
   }
 
+  /**
+   * Tear down.
+   *
+   * @throws IOException Signals that an I/O exception has occurred.
+   */
   @After
   public void tearDown() throws IOException {
     server.shutdown();
@@ -62,10 +88,8 @@ public class ProxyResourceTest {
   /**
    * Test send message.
    *
-   * @throws IOException
-   *           Signals that an I/O exception has occurred.
-   * @throws InterruptedException
-   *           the 4interrupted exception
+   * @throws IOException Signals that an I/O exception has occurred.
+   * @throws InterruptedException the 4interrupted exception
    */
   @Test
   public void testSendMessage() throws IOException, InterruptedException {
@@ -82,7 +106,7 @@ public class ProxyResourceTest {
     MessageRequest request = new MessageRequest.Builder().inputText(text).build();
     String payload = GsonSingleton.getGsonWithoutPrettyPrinting().toJson(request, MessageRequest.class);
 
-    InputStream inputStream = new ByteArrayInputStream(payload.getBytes());
+    InputStream inputStream = new ByteArrayInputStream(payload.getBytes("UTF-8"));
 
     Response jaxResponse = proxy.postMessage(WORKSPACE_ID, inputStream);
     MessageResponse serviceResponse = GsonSingleton.getGsonWithoutPrettyPrinting()
@@ -101,11 +125,26 @@ public class ProxyResourceTest {
     assertNotNull(mockRequest.getHeader(HttpHeaders.AUTHORIZATION));
   }
 
+  /**
+   * Load fixture.
+   *
+   * @param <T> the generic type
+   * @param filename the filename
+   * @param returnType the return type
+   * @return the t
+   * @throws FileNotFoundException the file not found exception
+   */
   public static <T> T loadFixture(String filename, Class<T> returnType) throws FileNotFoundException {
     String jsonString = getStringFromInputStream(new FileInputStream(filename));
     return new Gson().fromJson(jsonString, returnType);
   }
 
+  /**
+   * Gets the string from input stream.
+   *
+   * @param is the is
+   * @return the string from input stream
+   */
   public static String getStringFromInputStream(InputStream is) {
     BufferedReader br = null;
     final StringBuilder sb = new StringBuilder();
@@ -113,7 +152,7 @@ public class ProxyResourceTest {
     String line;
     try {
 
-      br = new BufferedReader(new InputStreamReader(is));
+      br = new BufferedReader(new InputStreamReader(is, "UTF-8"));
       while ((line = br.readLine()) != null) {
         sb.append(line);
       }
@@ -133,11 +172,9 @@ public class ProxyResourceTest {
   }
 
   /**
-   * Create a MockResponse with JSON content type and the object serialized to
-   * JSON as body.
+   * Create a MockResponse with JSON content type and the object serialized to JSON as body.
    *
-   * @param body
-   *          the body
+   * @param body the body
    * @return the mock response
    */
   protected static MockResponse jsonResponse(Object body) {
