+++ src/test/java/com/ibm/watson/apis/conversation_with_discovery/rest/SetupResourceTest.java
@@ -1,17 +1,16 @@
-/**
- * (C) Copyright IBM Corp. 2016. All Rights Reserved.
- * 
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License. You may obtain a copy of the License at
- * 
+/*
+ * Copyright 2015 IBM Corp. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
  * http://www.apache.org/licenses/LICENSE-2.0
- * 
- * Unless required by applicable law or agreed to in writing, software distributed under the License
- * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
- * or implied. See the License for the specific language governing permissions and limitations under
- * the License.
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
  */
-package com.ibm.watson.apis.conversation_enhanced.rest;
+package com.ibm.watson.apis.conversation_with_discovery.rest;
 
 import static org.junit.Assert.assertTrue;
 import static org.mockito.Mockito.mock;
@@ -27,27 +26,42 @@ import org.powermock.core.classloader.annotations.PrepareForTest;
 import org.powermock.modules.junit4.PowerMockRunner;
 
 import com.google.gson.JsonObject;
-import com.ibm.watson.apis.conversation_enhanced.listener.ServletContextListener;
+import com.ibm.watson.apis.conversation_enhanced.listener.AppServletContextListener;
 
 /**
- * Unit tests for the {@link SetupResource}
+ * Unit tests for the {@link SetupResource}.
  */
 
 @RunWith(PowerMockRunner.class)
 @PrepareForTest(SetupResource.class)
 public class SetupResourceTest {
-  SetupResource setupResource;
-  ServletContextListener mockServletContextListener = mock(ServletContextListener.class);
-  JsonObject config = new JsonObject();
-  String WORKSPACE_ID = "123456";
-  String EMPTY_WORKSPACE_ID = "";
 
+  /** The setup resource. */
+  SetupResource setupResource;
+
+  /** The mock servlet context listener. */
+  AppServletContextListener mockServletContextListener = mock(AppServletContextListener.class);
+
+  /** The config. */
+  JsonObject config = new JsonObject();
+
+  /** The workspace id. */
+  String WORKSPACE_ID = "123456";
+
+  /**
+   * Sets the up.
+   */
   @Before
   public void setUp() {
     setupResource = new SetupResource();
     PowerMockito.mockStatic(System.class);
   }
 
+  /**
+   * Should return ready state.
+   *
+   * @throws Exception the exception
+   */
   @Test
   public void shouldReturnReadyState() throws Exception {
 
@@ -57,7 +71,7 @@ public class SetupResourceTest {
     config.addProperty("setup_message", "all good");
 
     when(mockServletContextListener.getJsonConfig()).thenReturn(config);
-    PowerMockito.whenNew(ServletContextListener.class).withAnyArguments().thenReturn(mockServletContextListener);
+    PowerMockito.whenNew(AppServletContextListener.class).withAnyArguments().thenReturn(mockServletContextListener);
     when(System.getenv("WORKSPACE_ID")).thenReturn(WORKSPACE_ID);
     when(System.getenv("PASSWORD")).thenReturn("accdefghi");
     when(System.getenv("USERNAME")).thenReturn("abcd-efgh-ijkl");
@@ -69,6 +83,11 @@ public class SetupResourceTest {
     assertTrue(response.getEntity().toString().indexOf("\"setup_state\":\"ready\"") > 1);
   }
 
+  /**
+   * Should return not ready state.
+   *
+   * @throws Exception the exception
+   */
   @Test
   public void shouldReturnNotReadyState() throws Exception {
 
@@ -77,7 +96,7 @@ public class SetupResourceTest {
     config.addProperty("setup_state", "not_ready");
 
     when(mockServletContextListener.getJsonConfig()).thenReturn(config);
-    PowerMockito.whenNew(ServletContextListener.class).withAnyArguments().thenReturn(mockServletContextListener);
+    PowerMockito.whenNew(AppServletContextListener.class).withAnyArguments().thenReturn(mockServletContextListener);
     when(System.getenv("WORKSPACE_ID")).thenReturn(WORKSPACE_ID);
     when(System.getenv("PASSWORD")).thenReturn("accdefghi");
     when(System.getenv("USERNAME")).thenReturn("abcd-efgh-ijkl");
@@ -89,6 +108,11 @@ public class SetupResourceTest {
     assertTrue(response.getEntity().toString().indexOf("\"setup_state\":\"not_ready\"") > 1);
   }
 
+  /**
+   * Should return error state.
+   *
+   * @throws Exception the exception
+   */
   @Test
   public void shouldReturnErrorState() throws Exception {
 
@@ -101,7 +125,7 @@ public class SetupResourceTest {
     when(System.getenv("USERNAME")).thenReturn("abcd-efgh-ijkl");
 
     when(mockServletContextListener.getJsonConfig()).thenReturn(config);
-    PowerMockito.whenNew(ServletContextListener.class).withAnyArguments().thenReturn(mockServletContextListener);
+    PowerMockito.whenNew(AppServletContextListener.class).withAnyArguments().thenReturn(mockServletContextListener);
 
     // when
     Response response = setupResource.getConfig();
@@ -110,6 +134,11 @@ public class SetupResourceTest {
     assertTrue(response.getEntity().toString().indexOf("\"setup_message\":\"error\"") > 1);
   }
 
+  /**
+   * Should return error workspace id.
+   *
+   * @throws Exception the exception
+   */
   @Test
   public void shouldReturnErrorWorkspaceId() throws Exception {
 
@@ -121,7 +150,7 @@ public class SetupResourceTest {
     when(System.getenv("USERNAME")).thenReturn("abcd-efgh-ijkl");
 
     when(mockServletContextListener.getJsonConfig()).thenReturn(config);
-    PowerMockito.whenNew(ServletContextListener.class).withAnyArguments().thenReturn(mockServletContextListener);
+    PowerMockito.whenNew(AppServletContextListener.class).withAnyArguments().thenReturn(mockServletContextListener);
 
     // when
     Response response = setupResource.getConfig();
